pyproject.toml
[project]
name = "falcon"
version = "0.1.0"
description = "Interactive Brokers trading application"
requires-python = ">=3.10"
dependencies = [
    "ib_insync>=0.9.86",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
